Update README.md
Some react version compatibility is missing. Not sure how that should be noted. Maybe also add some link to https://github.com/timarney/react-app-rewired/issues/162 so people coming in with react or react-native apps that can't be rewired won't send in unwanted bug reports about things not working that are not supposed to be working.